Enable signed/unsigned mismatch warnings.  Allow GNU license
git-svn-id: svn+ssh://svn.corp.imvu.com/home/svnrepos/trunk/stable/Source/cal3d@120945 07c76cb3-cb09-0410-85de-c24d39f1912e

+++ plugins/cal3d_max_exporter/MaxMesh.cpp
@@ -825,7 +825,7 @@ CMaxMesh::numMorphChannels()
     return 0;
   }
   int activeCount = 0;
-  for( int i = 0; i < morpherModifier->chanBank.size(); i++ ) {
+  for(size_t i = 0; i < morpherModifier->chanBank.size(); i++) {
     morphChannel const & chanI = morpherModifier->chanBank[i];
     if( chanI.mActive ) {
       activeCount++;
